Prefix all modules with Ceres
`Format` and `String` are OCaml standard library modules,
and the other modules have generic enough names that
they are bound to cause issues too.

Ideally the Extraction plugin should use the full logical
path to produce the module name, so `theories/String.v`
extracts to `Ceres__String` for instance.
But for backwards compatibility, we're stuck with this kind
of manual mangling.